make a new migration to fix everyone on this alpha
