+++ NachoClient.Android/NachoClient.Android.csproj
@@ -499,6 +499,7 @@
     <Compile Include="NachoCore\Utils\ContactsBinningHelper.cs" />
     <Compile Include="NachoCore\Model\Migration\NcMigration35.cs" />
     <Compile Include="NachoCore\Utils\SearchHelper.cs" />
+    <Compile Include="NachoCore\Model\Migration\NcMigration36.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Assets\AboutAssets.txt" />
