+++ NachoClient.iOS/NachoClient.iOS.csproj
@@ -1580,6 +1580,9 @@
     <Compile Include="..\NachoClient.Android\NachoCore\Utils\SearchHelper.cs">
       <Link>NachoCore\Utils\SearchHelper.cs</Link>
     </Compile>
+    <Compile Include="..\NachoClient.Android\NachoCore\Model\Migration\NcMigration36.cs">
+      <Link>NachoCore\Model\Migration\NcMigration36.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <InterfaceDefinition Include="Resources\HomePageController.xib" />
